+++ app/src/androidTest/java/woowacourse/shopping/data/dao/CartDaoTest.kt
@@ -8,7 +8,7 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import woowacourse.shopping.data.database.ShoppingDatabase
-import woowacourse.shopping.data.entity.CartProductEntity
+import woowacourse.shopping.data.model.entity.CartProductEntity
 
 @Suppress("ktlint:standard:function-naming")
 class CartDaoTest {
