+++ app/src/androidTest/java/woowacourse/shopping/study/data/dao/CartDaoStudyTest.kt
@@ -12,7 +12,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.shopping.data.dao.CartDao
 import woowacourse.shopping.data.database.ShoppingDatabase
-import woowacourse.shopping.data.entity.CartProductEntity
+import woowacourse.shopping.data.model.entity.CartProductEntity
 
 @Suppress("ktlint:standard:function-naming")
 @RunWith(AndroidJUnit4::class)
