+++ app/src/main/java/woowacourse/shopping/data/api/ProductApi.kt
@@ -3,12 +3,12 @@ package woowacourse.shopping.data.api
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
-import woowacourse.shopping.data.dto.ProductDto
+import woowacourse.shopping.data.model.response.ProductResponse
 
 interface ProductApi {
     @GET("products")
     fun getProducts(
         @Query("lastId") lastId: Int,
         @Query("count") count: Int,
-    ): Call<List<ProductDto>>
+    ): Call<List<ProductResponse>>
 }
