+++ app/src/main/java/woowacourse/shopping/data/mapper/CartMapper.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.dto.CartProductDetailDto
+import woowacourse.shopping.data.model.response.CartProductDetailResponse
 import woowacourse.shopping.domain.model.CartProduct
 import woowacourse.shopping.domain.model.Product.Companion.EMPTY_PRODUCT
 
-fun CartProductDetailDto.toDomain(): CartProduct =
+fun CartProductDetailResponse.toDomain(): CartProduct =
     CartProduct(
         product = product?.toDomain() ?: EMPTY_PRODUCT,
         quantity = cartProduct.quantity,
