+++ app/src/main/java/woowacourse/shopping/data/mapper/HistoryMapper.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.dto.HistoryProductDto
+import woowacourse.shopping.data.model.response.HistoryProductResponse
 import woowacourse.shopping.domain.model.HistoryProduct
 import woowacourse.shopping.domain.model.HistoryProduct.Companion.EMPTY_HISTORY_PRODUCT
 
-fun HistoryProductDto.toDomain(): HistoryProduct =
+fun HistoryProductResponse.toDomain(): HistoryProduct =
     if (product != null) {
         HistoryProduct(
             productId = product.id,
