+++ app/src/main/java/woowacourse/shopping/data/model/entity/CartProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.entity
+package woowacourse.shopping.data.model.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
