+++ app/src/main/java/woowacourse/shopping/data/model/response/HistoryProductResponse.kt
@@ -1,11 +1,11 @@
-package woowacourse.shopping.data.dto
+package woowacourse.shopping.data.model.response
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import woowacourse.shopping.data.entity.HistoryProductEntity
-import woowacourse.shopping.data.entity.ProductEntity
+import woowacourse.shopping.data.model.entity.HistoryProductEntity
+import woowacourse.shopping.data.model.entity.ProductEntity
 
-data class HistoryProductDto(
+data class HistoryProductResponse(
     @Embedded val historyProduct: HistoryProductEntity,
     @Relation(
         parentColumn = "productId",
