+++ app/src/main/java/woowacourse/shopping/data/repository/CartRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.data.dao.CartDao
-import woowacourse.shopping.data.entity.CartProductEntity
 import woowacourse.shopping.data.mapper.toDomain
+import woowacourse.shopping.data.model.entity.CartProductEntity
 import woowacourse.shopping.domain.model.CartProduct
 import woowacourse.shopping.domain.repository.CartRepository
