+++ app/src/main/java/woowacourse/shopping/data/repository/HistoryRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.data.dao.HistoryDao
-import woowacourse.shopping.data.entity.HistoryProductEntity
 import woowacourse.shopping.data.mapper.toDomain
+import woowacourse.shopping.data.model.entity.HistoryProductEntity
 import woowacourse.shopping.domain.model.HistoryProduct
 import woowacourse.shopping.domain.repository.HistoryRepository
